validation/requirements.txt
numpy==1.25.2
opencv_python==4.10.0.82
Pillow==10.4.0
scikit-image
torch
torchvision
fastapi
uvicorn
transformers
trimesh
open-clip-torch==2.23.0
timm==0.9.8